+++ src_test/zh_HK/test_zh_HK.ml
@@ -1,0 +1,15 @@
+open OUnit2
+
+(*
+let _ =
+  let foo = 1 in
+  if x = 1 then "x"
+  else "y"
+*)
+
+设 _ =
+  设 x = 1 在
+  如果 x = 1 那么 "x"
+  否则 "y"
+
+let suite = "zh_HK" >::: []