+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.1
+---
+
+  * OCaml 4.04 compatibility.
+  * Localization of keywords to zh_HK.
+
 1.0
 ---